+++ raster_loader/io/snowflake.py
@@ -33,21 +33,24 @@ else:
 
 
 class SnowflakeConnection(DataWarehouseConnection):
-    def __init__(self, username, password, account, database, schema, token, role):
+    def __init__(
+        self,
+        username,
+        password,
+        account,
+        database,
+        schema,
+        token,
+        private_key_path,
+        private_key_passphrase,
+        role,
+        warehouse,
+    ):
         if not _has_snowflake:
             import_error_snowflake()
 
         # TODO: Write a proper static factory for this
-        if token is None:
-            self.client = snowflake.connector.connect(
-                user=username,
-                password=password,
-                account=account,
-                database=database.upper(),
-                schema=schema.upper(),
-                role=role.upper() if role is not None else None,
-            )
-        else:
+        if token is not None:
             self.client = snowflake.connector.connect(
                 authenticator="oauth",
                 token=token,
@@ -55,6 +58,29 @@ class SnowflakeConnection(DataWarehouseConnection):
                 database=database.upper(),
                 schema=schema.upper(),
                 role=role.upper() if role is not None else None,
+                warehouse=warehouse,
+            )
+        elif private_key_path is not None:
+            self.client = snowflake.connector.connect(
+                authenticator="snowflake_jwt",
+                user=username,
+                private_key_file=private_key_path,
+                private_key_file_pwd=private_key_passphrase,
+                account=account,
+                database=database.upper(),
+                schema=schema.upper(),
+                role=role.upper() if role is not None else None,
+                warehouse=warehouse,
+            )
+        else:
+            self.client = snowflake.connector.connect(
+                user=username,
+                password=password,
+                account=account,
+                database=database.upper(),
+                schema=schema.upper(),
+                role=role.upper() if role is not None else None,
+                warehouse=warehouse,
             )
 
     def band_rename_function(self, band_name: str):
@@ -179,8 +205,12 @@ class SnowflakeConnection(DataWarehouseConnection):
         append: bool = False,
         cleanup_on_failure: bool = False,
         exact_stats: bool = False,
-        all_stats: bool = False,
+        basic_stats: bool = False,
+        compress: bool = False,
+        compression_level: int = 6,
     ) -> bool:
+        """Write a raster file to a Snowflake table."""
+
         def band_rename_function(x):
             return x.upper()
 
@@ -207,18 +237,22 @@ class SnowflakeConnection(DataWarehouseConnection):
                     exit()
 
             metadata = rasterio_metadata(
-                file_path, bands_info, band_rename_function, exact_stats, all_stats
+                file_path, bands_info, band_rename_function, exact_stats, basic_stats
             )
 
             overviews_records_gen = rasterio_overview_to_records(
                 file_path,
                 band_rename_function,
                 bands_info,
+                compress=compress,
+                compression_level=compression_level,
             )
             windows_records_gen = rasterio_windows_to_records(
                 file_path,
                 band_rename_function,
                 bands_info,
+                compress=compress,
+                compression_level=compression_level,
             )
 
             records_gen = chain(overviews_records_gen, windows_records_gen)